Handled every warning message while compiling.

+++ src/DemandProfile.h
@@ -25,13 +25,13 @@ public:
 	void reset();
 	void getDemandInfo(double* cX, double* cY, double* laPerPixel, double* loPerPixel, string* timeStart, string* timeEnd);
 	void getRange(double* minLati, double* minLong, double* maxLati, double* maxLong);
-	void generateDemandVector(vector<float> &demandRNPs, double startingAngle, double endingAngle, float nmilesPerPixel);
+	void generateDemandVector(vector<double> &demandRNPs, double startingAngle, double endingAngle, double nmilesPerPixel);
 	
 	// most x,y pairs we are talking about in this class are actually lati/long pairs
 private:
 	vector<double> xCoors;	/**< the x coordinates of the aircraft (latitudes) */
 	vector<double> yCoors;	/**< the y coordinates of the aircraft (longitudes) */
-	vector<float> rnp; /**< the rnp of each aircraft */
+	vector<double> rnp; /**< the rnp of each aircraft */
 	string timeStart; /**< Simulation starting time */
         string timeEnd;   /**< Simulation ending time */
 	double centerX;	// the x coordinate of the airport(latitude)
@@ -40,7 +40,7 @@ private:
 	double longPerPixel; // conversion from longitude to pixel in y direction
 	double altPerPixel;  
         double minX, minY, maxX, maxY;	// the ranges of the lati/long values
-	int numDemands;		// the number of demands that are listed in the input file
+	unsigned int numDemands;		// the number of demands that are listed in the input file
 private:
 	bool handleInputData();	 // after reading in, deal with the input
         // test if we are going out of the range of the file, meaning file format error
